Skip Sparkle tools linking and framework stripping in Debug builds to reduce build times
- Speed up iteration time for Debug builds by making some build phases no-ops, resulting in less follow-up work needing to be done in later build phases
- Add early exit conditions when building in Debug configuration
- Avoid unnecessary re-copying and re-code-signing of framework contents
- Keep existing symlinks and framework state if already present

+++ Configurations/link-tools.sh
@@ -12,5 +12,13 @@ fi
 
 FRAMEWORK_PATH="${TARGET_BUILD_DIR}"/"${FULL_PRODUCT_NAME}"
 
+# Skip if configuration is debug and the targets exist already; together with similar changes in Sparkle's other build
+# scripts, this can help speed up the iteration time for debug builds by avoiding having to re-copy and re-code-sign the
+# framework and its contents upon every build.
+if [[ "$CONFIGURATION" == "Debug" && -e "${FRAMEWORK_PATH}"/"${SPARKLE_RELAUNCH_TOOL_NAME}" && -e "${FRAMEWORK_PATH}"/"${SPARKLE_INSTALLER_PROGRESS_TOOL_NAME}".app ]]; then
+    echo "Skipping linking of (already-linked) Sparkle tools for Debug configuration to reduce build iteration time"
+    exit 0
+fi
+
 ln -h -f -s "Versions/Current/""${SPARKLE_RELAUNCH_TOOL_NAME}" "${FRAMEWORK_PATH}"/"${SPARKLE_RELAUNCH_TOOL_NAME}"
 ln -h -f -s "Versions/Current/""${SPARKLE_INSTALLER_PROGRESS_TOOL_NAME}".app "${FRAMEWORK_PATH}"/"${SPARKLE_INSTALLER_PROGRESS_TOOL_NAME}".app
